requirements.txt
gunicorn==19.9.0
falcon==1.4.1
sqlalchemy==1.2.12
boto3==1.9.9
psycopg2==2.7.5
requests==2.19.1
pyyaml==3.13
